Totally change the way orange-cat works.
To implement a feature request #5, the way it works has been completely
changed. Now orange-cat doesn't open any channel or socket by default,
but opens a data channel and Websocket when a Markdown file is requested
by a client.

If it's not a Markdown file, it will be served as a static file.

MdChan's been removed. Now MdConverter is a global variable.